Renamed the different module list item to include "ModuleListItem", so it won't conflict with other list item names.
