--- Source/Clients/NostalgicPlayer/MainWindow/ListItem/SingleFileListItem.cs
@@ -1,60 +1,0 @@
-﻿/******************************************************************************/
-/* This source, or parts thereof, may be used in any software as long the     */
-/* license of NostalgicPlayer is keep. See the LICENSE file for more          */
-/* information.                                                               */
-/******************************************************************************/
-using System.IO;
-using Polycode.NostalgicPlayer.PlayerLibrary.Agent;
-using Polycode.NostalgicPlayer.PlayerLibrary.Loaders;
-
-namespace Polycode.NostalgicPlayer.Client.GuiPlayer.MainWindow.ListItem
-{
-	/// <summary>
-	/// This class holds a list item pointing to a file
-	/// </summary>
-	public class SingleFileListItem : IModuleListItem
-	{
-		/********************************************************************/
-		/// <summary>
-		/// Constructor
-		/// </summary>
-		/********************************************************************/
-		public SingleFileListItem(string fullFileName)
-		{
-			Source = fullFileName;
-		}
-
-		#region IModuleListItem implementation
-		/********************************************************************/
-		/// <summary>
-		/// Return the name which is shown in the list
-		/// </summary>
-		/********************************************************************/
-		public string DisplayName => Path.GetFileName(Source);
-
-
-
-		/********************************************************************/
-		/// <summary>
-		/// Return the full path to the file
-		/// </summary>
-		/********************************************************************/
-		public string Source
-		{
-			get;
-		}
-
-
-
-		/********************************************************************/
-		/// <summary>
-		/// Create the loader to use to load this item
-		/// </summary>
-		/********************************************************************/
-		public LoaderBase CreateLoader(Manager agentManager)
-		{
-			return new Loader(agentManager);
-		}
-		#endregion
-	}
-}
